svn merge -r 29045:29046 svn://chef/work/src . (Comment: #8788)
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/branch/prod/release-9.0/work/src@29047 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ jnetcap/impl/com/untangle/jnetcap/Endpoints.java
@@ -27,5 +27,5 @@ public interface Endpoints
     /**
      * Retrieve a unique interface identifier, or 0 if the interface is unknown
      */
-    public byte interfaceId();
+    public int interfaceId();
 }
